fix(windows): prevent infinite run. Fixes #11810 (#11993)
